requirements.txt
asgiref==3.8.1
certifi==2024.12.14
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
Django==5.1.5
django-cors-headers==4.6.0
django-rest-knox==5.0.2
djangorestframework==3.15.2
dynaconf==3.2.6
hvac==2.3.0
idna==3.10
knox==0.1.14
loguru==0.7.3
pillow==11.1.0
requests==2.32.3
shortuuid==1.0.13
sqlparse==0.5.3
tzdata==2024.2
urllib3==2.3.0
validators==0.34.0
win32_setctime==1.2.0
